% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_addParentAphiaIDs.R
\name{addParentAphiaIDs}
\alias{addParentAphiaIDs}
\title{Add Parent AphiaID}
\usage{
addParentAphiaIDs(dataset)
}
\arguments{
\item{dataset}{data frame that has the column "AphiaID".}
}
\description{
This function allows you add Parent records to your data frame and outputs missing parents according to their status.
}
\details{
This function depends on 'missingParentAphiaIDs' and 'updateParentAphiaIDs' functions.
This function differs from those because of the outputs. Four outputs will be generated to identify the Parent Aphia ID records that have been added by their status.

Outputs:

dataset - input dataset with missing parents added in. If there is no "Parent_AphiaID" column the function will run the 'updateParentAphiaID' function.

accepted_missing_parents - missing parent records that have accepted valid AphiaI IDs.

alt_rep_missing_parents - missing parent records with a valid status of 'alternate representation'.

other_missing_parents - all other missing parent records (not accepted or an alternate representation).
}
\examples{
data(marineKingdoms)
#run add parents function
u <- addParentAphiaIDs(marineKingdoms)

#check for missing parents as individual records
missingParentAphiaIDs(u)
# Error in missingParentAphiaIDs(u) :
#  Some Parent AphiaIDs are NA. Re- run 'updateParentAphiaIDs' function.

#run update parent IDs function
v <- updateParentAphiaIDs(u)

#check for missing parents as individual records
missingParentAphiaIDs(v)
#[1] "All parent aphiaIDs are also individual records."

}
\keyword{AphiaID,}
\keyword{ParentAphiaID}
\keyword{WoRMS,}
