% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_addValidAphiaIDs.R
\name{addValidAphiaIDs}
\alias{addValidAphiaIDs}
\title{Add Valid AphiaID}
\usage{
addValidAphiaIDs(dataset)
}
\arguments{
\item{dataset}{data frame that has the column "AphiaID".}
}
\description{
This function allows you add Valid Aphia ID records to your data frame and outputs changes made to the dataframe.
}
\details{
This function depends on the 'missingValidAphiaIDs' function.
This function differs from those as because of the outputs. Outputs will be generated to know the records that have been added or amended.

Outputs:
dataset - input dataset with AphiaID and validAphiaIDs.
added records - this will presented in the R Console.
validAphiaIDUpdates - saved as csv to your working directory. This is only created when an valid aphia ID is not accepted but there is an accepted valid aphia ID.
}
\examples{
data(marineKingdoms)
#run add valid IDs function
x <- addValidAphiaIDs(marineKingdoms)
#Added Aphia ID records: 6
#1 new records have been added.

# Aphia ID 6 is a valid Aphia ID but is not in the dataset as an individual record.

#Run to check if there any more missing valid AphiaID records
missingValidAphiaIDs(x)
#[1] "All valid aphiaIDs are also individual records"


}
\keyword{AphiaID,}
\keyword{ParentAphiaID}
\keyword{WoRMS,}
