% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_createShortcodes.R
\name{createShortcodes}
\alias{createShortcodes}
\title{Create 3 + 3 letter short codes from a string}
\usage{
createShortcodes(string)
}
\arguments{
\item{string}{The string you would like to create a code for.}
}
\description{
This function will take a string and return the first three letters of the first word and first three letters of the second word.
}
\examples{
# Enter a string
createShortcodes("Salmo salar")
#[1] "Salsal"
}
\keyword{abbreviations}
\keyword{shortcodes,}
\keyword{string,}
