% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_getParentID.R
\name{getParentID}
\alias{getParentID}
\title{Get Parent AphiaID}
\usage{
getParentID(inputAphiaID)
}
\arguments{
\item{inputAphiaID}{The Aphia ID of the taxon that you want the parent Aphia ID from.}
}
\description{
This function allows you retrieve the parent of a given Aphia ID.
}
\examples{
# Enter aphia ID - Animalia (2)
getParentID(2)
#[1] 1
}
\keyword{AphiaID,}
\keyword{WoRMS}
