% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineKingdoms.R
\docType{data}
\name{marineKingdoms}
\alias{marineKingdoms}
\title{Data from marineKingdoms}
\format{A data frame with 13 rows and 25 variables:}
\source{
WoRMS
}
\usage{
data(marineKingdoms)
}
\description{
This is data from World Register of Marine Species (WoRMS). The data is of all the children of 'Biota' without Archaea but including its children.
}
\keyword{datasets}
