% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_missingParentAphiaIDs.R
\name{missingParentAphiaIDs}
\alias{missingParentAphiaIDs}
\title{Find records with missing Parent records}
\usage{
missingParentAphiaIDs(dataset)
}
\arguments{
\item{dataset}{data frame that has the column "AphiaID" and "Parent.AphiaID.}
}
\description{
This function allows you to identify missing Parent records from your data frame.
}
\examples{
data(marineKingdoms)

#run 'updateParentAphiaIDs' function
x <- updateParentAphiaIDs(marineKingdoms)

#Run 'missingParentAphiaIDs' function to check if any parent records are missing from the dataset.
missingParentAphiaIDs(x)
#[1] 1 8

}
\keyword{AphiaID,}
\keyword{Parent}
\keyword{WoRMS,}
