% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_missingValidAphiaIDs.R
\name{missingValidAphiaIDs}
\alias{missingValidAphiaIDs}
\title{Find aphia ID records with missing valid records in the dataset}
\usage{
missingValidAphiaIDs(dataset)
}
\arguments{
\item{dataset}{data frame that has the column "AphiaID" and "valid_AphiaID".}
}
\description{
This function identifies if the valid aphia IDs are entered as individual records
}
\details{
To use this function the dataset must have an "AphiaID" column and a "valid_AphiaID" column.
}
\examples{
data(marineKingdoms)

#Run missingValidAphiaIDs function to check if any Valid Aphia ID records are missing from the dataset.
missingValidAphiaIDs(marineKingdoms)
#[1] 6
# Aphia ID 6 is a valid AphiaID for a record but is missing from the dataset. Use 'addValidAphiaIDs' function to enter this record.

}
\keyword{AphiaID,}
\keyword{Parent}
\keyword{WoRMS,}
