% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_MR_GIS_Species_shp.R
\name{MR_GIS_Species}
\alias{MR_GIS_Species}
\title{Marine Recorder GIS species extract}
\usage{
MR_GIS_Species(snapshot_filepath, snapshot_version)
}
\arguments{
\item{snapshot_filepath}{The file path to the the snapshot. Please enter this filepath in quotations.}

\item{snapshot_version}{The version of the MR snapshot you are using. Please enter these in quotations e.g. "MRv5_Public_20161205"}
}
\description{
This function creates a shapefile of the species records found in a marine recorder snapshot. It saves the output to your working directory.
}
\details{
For this function to work you will need to use a 32-bit version R.
This can be done by clicking on the 'Tools' tab -> 'Global Options' -> 'General' -> 'R Version' click 'Change' and select the 32-bit version of R.
The function will return a shapefile saved to your working directory.
The saved shapefile will follow this format ("C", Sys.Date(), "_Species_", snapshot_version) e.g. 'C2018-05-05_Species_MRv5_Public_20161205.shp')
}
\examples{
MR_GIS_Species()
}
\keyword{GIS,}
\keyword{Marine}
\keyword{Recorder,}
\keyword{shapefile,}
\keyword{species}
