% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_MRtable_compare.R
\name{newSurveys}
\alias{newSurveys}
\title{Find new surveys in the snapshot}
\usage{
newSurveys(DSN1, tablename1, DSN2, tablename2, primaryKey)
}
\arguments{
\item{DSN1}{- file location of database 1}

\item{tablename1}{- name of table in quotations for comparison in database 1}

\item{DSN2}{- file location of database 2}

\item{tablename2}{- name of table in quotations for comparison in database 2}

\item{primaryKey}{- column name of the primary key (the column used to compare the tables)}
}
\description{
This function will identify new surveys that have been added into the new snapshot.
}
\examples{

#find and paste new snapshot filepath
filepath_new_snapshot <- "SnapshotDatav51_Public_20180524.mdb"

#find and paste old snapshot filepath
filepath_old_snapshot <- "SnapshotDatav51_Public_20180131.mdb"

#run the function on the 'Survey' table where the primary key for the table is 'Survey_Key'
surveysAdded <- newSurveys(DSN1 = filepath_new_snapshot, tablename1 = "Survey", DSN2 = filepath_old_snapshot, tablename2 = "Survey", primaryKey = "Survey_Key" )

}
\keyword{RODBC,}
\keyword{snapshot,}
\keyword{surveys}
