% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_lastKeyUpdate.R
\name{updateLastKeyNBN}
\alias{updateLastKeyNBN}
\title{update the Last Key table}
\usage{
updateLastKeyNBN(NBNfilepath)
}
\arguments{
\item{NBNfilepath}{filepath of Marine Recorder NBN}
}
\description{
This function will update the Last Key table in the NBN data file.
}
\details{
Only use this function if you encounter "Run-Time Error '3022': The Changes you Requested to the Table were not Successful", when attempting to enter data into Marine Recorder.

When entering data into Marine Recorder make sure that you have the debug = 1 (this can be found in the .ini file where the Marine Recorder Application was installed).

Please ensure that you are working on a copy of the NBN data file and not the original.

The function will automatically update the NBN and return a list of the changes that have been made to the NBN.
}
\examples{

## not run
#find and paste NBN filepath
filepathNBN<- ""

#run the function
lastKeyChanges <- updateLastKeyNBN(filepathNBN)

## End (not run)
}
\keyword{NBN,}
\keyword{RODBC,}
\keyword{key}
\keyword{last}
