% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_updateParentAphiaIDs.R
\name{updateParentAphiaIDs}
\alias{updateParentAphiaIDs}
\title{Update or Add Parent Aphia IDs}
\usage{
updateParentAphiaIDs(dataset)
}
\arguments{
\item{dataset}{data frame that has the column "AphiaID".}
}
\description{
This function allows you add Parent records to your data frame. If there is no column called "Parent.AphidID" it will generate one and populate it.
}
\examples{
data(marineKingdoms)

#run 'updateParentAphiaIDs' function. It add in the column 'Parent.AphiaID' and populates the column using the  'getParentID' function.
x <- updateParentAphiaIDs(marineKingdoms)

#view the outputs
View(x)

}
\keyword{AphiaID,}
\keyword{WoRMS}
